+++ rust/src/snmp/detect.rs
@@ -22,7 +22,7 @@ use crate::snmp::snmp::SNMPTransaction;
 #[no_mangle]
 pub unsafe extern "C" fn rs_snmp_tx_get_version(tx: &mut SNMPTransaction, version: *mut u32) {
     debug_assert!(tx.version != 0, "SNMP version is 0");
-    *version = tx.version as u32;
+    *version = tx.version;
 }
 
 #[no_mangle]
@@ -42,7 +42,7 @@ pub unsafe extern "C" fn rs_snmp_tx_get_community(
 pub unsafe extern "C" fn rs_snmp_tx_get_pdu_type(tx: &mut SNMPTransaction, pdu_type: *mut u32) {
     match tx.info {
         Some(ref info) => {
-            *pdu_type = info.pdu_type.0 as u32;
+            *pdu_type = info.pdu_type.0;
         }
         None => {
             *pdu_type = 0xffffffff;
